Get Ready For Custom Seed And Daily run(1)

+++ Arrrknights/core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/Bones.java
@@ -52,8 +52,8 @@ public class Bones {
 
 		depth = Dungeon.depth;
 
-		//heroes which have won the game, who die far above their farthest depth, or who are challenged drop no bones.
-		if (Statistics.amuletObtained || (Statistics.deepestFloor - 5) >= depth || Dungeon.challenges > 0 || Dungeon.isChallenged(Challenges.TEST)) {//change from budding
+		//heroes drop no bones if they have the amulet, die far above their farthest depth, are challenged, or are playing with a custom seed.
+		if (Statistics.amuletObtained || (Statistics.deepestFloor - 5) >= depth || Dungeon.challenges > 0 || !Badges.is_not_test || !Dungeon.customSeedText.isEmpty()) {//change from budding
 			depth = -1;
 			return;
 		}
@@ -142,7 +142,7 @@ public class Bones {
 
 		} else {
 			//heroes who are challenged cannot find bones
-			if (depth == Dungeon.depth && Dungeon.challenges == 0) {
+			if (depth == Dungeon.depth && Dungeon.challenges == 0 && Dungeon.customSeedText.isEmpty() && Badges.is_not_test) {
 				FileUtils.deleteFile( BONES_FILE );
 				depth = 0;
